Merge  Fix compilation with GCC 6
GCC 6 defaults to a newer C++ standard version.  C++11 introduced a new
overload for push_back so it is now sometimes necessary to specify which
overload is required.  C++11 also introduced std::bind so we need to
specify the namespace when using boost::lambda::bind.

by  jascrain committed on 1 May 2016
commit b5bd0370a569298d13b258ea4e57e48f5b03ecbf
from master branch